setup/ruben/requirements.txt
groq>=0.9.0
ollama>=0.3.0
clearml>=1.16.0
python-dotenv>=1.0.0
